# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/form/add": {
            "post": {
                "description": "新增表单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "新增表单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FormInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/all": {
            "get": {
                "description": "查询所有表单管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "查询所有表单管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetAllFormResponse"
                        }
                    }
                }
            }
        },
        "/api/form/copy": {
            "post": {
                "description": "复制表单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "复制表单",
                "parameters": [
                    {
                        "description": "Copy Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/delete": {
            "delete": {
                "description": "删除表单管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "删除表单管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/detail": {
            "get": {
                "description": "查询明细",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "查询明细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetFormDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/form/export": {
            "get": {
                "description": "导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "导出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "formIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "formSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "表单名称",
                        "name": "pageName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/import": {
            "post": {
                "description": "导入",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "导入",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer+空格+Token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/publish": {
            "post": {
                "description": "发布表单版本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "发布表单版本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FormVersion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/query": {
            "get": {
                "description": "分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "分页查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "formIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "formSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "表单名称",
                        "name": "pageName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.QueryFormResponse"
                        }
                    }
                }
            }
        },
        "/api/form/schema": {
            "put": {
                "description": "更新表单Schema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "更新表单Schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FormInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/switch": {
            "post": {
                "description": "切换表单版本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "切换表单版本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version ID",
                        "name": "versionID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/update": {
            "put": {
                "description": "更新表单管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "更新表单管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Form",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FormInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/form/version/detail": {
            "get": {
                "description": "查询明细",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "表单管理"
                ],
                "summary": "查询明细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetVersionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "curd.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "curd.DelRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
...
        "curd.FormInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isMust": {
                    "description": "系统必须要有的数据",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "pageName": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "schema": {
                    "type": "string"
                },
                "subform": {
                    "type": "boolean"
                },
                "tenantID": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "versions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.FormVersion"
                    }
                }
            }
        },
...
    }
}